DevNet: sync approved-sv-id-values with ledger state
Following weights shift from DA-1 to DA-2 to prepare for DA-1 reonboarding.

Note that the current total weight for DA is wrong atm due to an operator error. Fixing/avoiding that would have delayed our timeline for reonboarding DA-1, so we decided to ignore. This PR now just syncs up to the current state on ledger.

[allow-total-reward-weight-change]

Signed-off-by: Martin Florian <martin.florian@digitalasset.com>